Fix TAS run

+++ include/global.h
@@ -18,7 +18,7 @@
 
 #if !PLATFORM_GBA
 #ifdef _WIN32
-void *Platform_malloc(int numBytes);
+void *Platform_malloc(size_t numBytes);
 void Platform_free(void *ptr);
 #define malloc(numBytes)    Platform_malloc(numBytes)
 #define calloc(count, size) Platform_malloc(count *size)
